Refactor component imports
