+++ src/app/components/common/bottomMessageBar/BottomMessageBar.js
@@ -1,7 +1,7 @@
 "use client";
 
 import {useContext} from 'react';
-import {BottomMessageBarContext} from '../context/bottomMessageBar/BottomMessageBarContext';
+import {BottomMessageBarContext} from '../../../context/bottomMessageBar/BottomMessageBarContext';
 
 
 export function BottomMessageBar() {
