+++ src/app/resume/page.js
@@ -1,5 +1,5 @@
-import Container from "../components/Container";
-import ResumeDownloader from "../components/ResumeDownloader";
+import Container from "../components/common/container/Container";
+import ResumeDownloader from "../components/resumeComponents/ResumeDownloader";
 
 export default function Resume() {
     return (
