+++ src/app/sections/Skills.js
@@ -1,5 +1,5 @@
 import Image from 'next/image'
-import Container from "../components/Container";
+import Container from "../components/common/container/Container";
 
 const skills = [
     {
